+++ branches/win-version/Sources/MWindowImpl.h
@@ -42,7 +42,7 @@ class MWindowImpl
 
 	virtual void		ScrollRect(MRect inRect, int32 inDeltaH, int32 inDeltaV) = 0;
 	
-	virtual bool		GetMouse(int32& outX, int32& outY, unsigned long& outModifiers) = 0;
+	virtual bool		GetMouse(int32& outX, int32& outY, uint32& outModifiers) = 0;
 	virtual bool		WaitMouseMoved(int32 inX, int32 inY) = 0;
 
 	virtual void		ConvertToScreen(int32& ioX, int32& ioY) const = 0;
